pyproject.toml
[tool.poetry]
name = "story-llm"
version = "0.1.0"
description = "A backend service to evaluate product ideas using AI agents."
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
motor = "*"
pydantic-ai = "*"
logfire = {extras = ["fastapi"], version = "^4.16.0"}
opik = "*"
python-dotenv = "*"
openai = "*"
email-validator = "*"
passlib = {extras = ["bcrypt"], version = "*"}
python-jose = {extras = ["cryptography"], version = "*"}
python-multipart = "*"
opentelemetry-instrumentation-fastapi = "^0.60b1"
bcrypt = "^5.0.0"
argon2-cffi = "^25.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
httpx = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
